requirements.txt
crewai[tools,agentops]==0.102.0
python-dotenv==1.0.1
tavily-python
scrapegraph-py
